// src/graphql/generated/version.generated.tsx
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  DateTime: any;
  Upload: any;
};
// ...
export type BioDto = {
  __typename?: "BioDto";
  status?: Maybe<Scalars["String"]>;
  time?: Maybe<Scalars["Float"]>;
};
// ...
export type ContactDetailsDto = {
  __typename?: "ContactDetailsDto";
  additional?: Maybe<Scalars["String"]>;
  address?: Maybe<Scalars["String"]>;
  blocked?: Maybe<Scalars["Boolean"]>;
  city?: Maybe<Scalars["String"]>;
  country?: Maybe<Scalars["String"]>;
  dob?: Maybe<Scalars["DateTime"]>;
  email?: Maybe<Scalars["String"]>;
  firstName?: Maybe<Scalars["String"]>;
  gender?: Maybe<Scalars["String"]>;
  hasComon?: Maybe<Scalars["Boolean"]>;
  hasInvited?: Maybe<Scalars["Boolean"]>;
  invitedAt?: Maybe<Scalars["Float"]>;
  lastName?: Maybe<Scalars["String"]>;
  lastSeen?: Maybe<Scalars["Float"]>;
  localId: Scalars["String"];
  originalPhone: Scalars["String"];
  phone: Scalars["String"];
  prefix?: Maybe<Scalars["String"]>;
  region?: Maybe<Scalars["String"]>;
  status?: Maybe<Scalars["String"]>;
  street?: Maybe<Scalars["String"]>;
  suffix?: Maybe<Scalars["String"]>;
  userId?: Maybe<UserDetailsDto>;
  website?: Maybe<Scalars["String"]>;
};
// ...
export type UserDetailsDto = {
  __typename?: "UserDetailsDto";
  _id: Scalars["String"];
  bio: BioDto;
  firstName: Scalars["String"];
  lastName: Scalars["String"];
  lastSeen?: Maybe<Scalars["Float"]>;
  phone: Scalars["String"];
  profile_img: Scalars["String"];
  status: Scalars["String"];
};